# docs/apidocs/localapidoc.yaml
swagger: '2.0'
info:
  description: This is an api doc for `seda configurator` in sap cloud foundry.
  version: 0.0.1
  title: Seda Configurator
  termsOfService: ''
  contact:
    email: gokulraj.ramdass@sap.com
  license:
    name: MIT
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: 'localhost:3000'
basePath: /api/v1
...
schemes:
  - http
paths:
...
  '/apps/{appid}/intents/all':
    get:
      tags:
        - intents
      summary: get all intents
      description: get all intents for the given application ID.
      operationId: getallintents
      produces:
        - application/json
      parameters:
        - in: path
          name: appid
          description: id of the application
          required: true
          type: string
      responses:
        '200':
          description: Success data fetched !
        '405':
          description: Invalid Input or App ID
        '500':
          description: Error Occurred during processing
...